Move files to within geotrellis.contrib.vlm.avro package
